requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
openai>=1.30.0
chromadb>=0.5.0
langsmith>=0.3.0
pydantic>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
httpx>=0.27.0
pytest>=8.0.0
langchain-chroma>=0.1.0
gradio>=4.0.0
snowballstemmer>=3.0.1
